api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pydantic==2.5.0
openai==1.54.0
httpx==0.27.2
PyJWT==2.8.0
pytest==7.4.3
